OAStackView can be embedded in extensions
Sets the "Allow App Extension API only" checkbox
for the OAStackView build target. This setting is
required to use OAStackView in an app extension, as it causes
the compiler to enforce that OAStackView does
not use API that is disallowed within app extensions. It
should have no effect on the use of OAStackView outside
of extensions.